implemented AntDecorator
